namespace: use ZulipAPI namespace.
