+++ SampleApp/UserControls/UCStreams.cs
@@ -1,5 +1,5 @@
 using System.Windows.Forms;
-using ZulipNetCore;
+using ZulipAPI;
 
 namespace SampleApp.UserControls {
     public partial class UCStreams : UserControl {
