+++ Tests/UnitTests.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NetCore\NetCore.csproj" />
+    <ProjectReference Include="..\ZulipAPI\ZulipAPI.csproj" />
   </ItemGroup>
 
 </Project>
