+++ ZulipAPI/BaseClasses/EndPointSendMessage.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
 
     public abstract class EndPointSendMessage {
