+++ ZulipAPI/Exceptions/InvalidZulipRCFileException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
     internal class InvalidZulipRCFileException : Exception {
 
         public InvalidZulipRCFileException() : base() {
