+++ ZulipAPI/Interfaces/IParseResponse.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace ZulipNetCore.Interfaces {
+namespace ZulipAPI.Interfaces {
 
     internal interface IParseResponse {
