+++ ZulipAPI/Messages/MessageCollection.cs
@@ -1,26 +1,23 @@
 using System.Collections;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
 
-    public class StreamCollection : CollectionBase, IEnumerable, IEnumerator {
+    public class MessageCollection : CollectionBase, IEnumerable, IEnumerator {
 
         private int index = -1;
 
-        public StreamCollection() {
+        public MessageCollection() {
             this.index = -1;
         }
 
-        public void Add(Stream stream) {
-            if (stream != null) {
-                this.List.Add(stream);
+        public void Add(Message message) {
+            if (message != null) {
+                this.List.Add(message);
             }
         }
 
-        public void Remove(Stream stream) {
-            this.List.Remove(stream);
+        public void Remove(Message message) {
+            this.List.Remove(message);
         }
 
         IEnumerator IEnumerable.GetEnumerator() {
@@ -41,6 +38,5 @@ namespace ZulipNetCore {
         public void Reset() {
             this.index = -1;
         }
-
     }
 }
