+++ ZulipAPI/Streams/Stream.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
 
     public class Stream {
