+++ ZulipAPI/Streams/Streams.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
 
     public class Streams : EndPointBase {
