+++ ZulipAPI/Users/BotTypes.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace ZulipNetCore {
+namespace ZulipAPI {
 
     public enum BotTypes {
         GenericBot =0,
