+++ ZulipCSharp.sln
@@ -6,7 +6,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleApp", "SampleApp\Samp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZulipClient", "ZulipClient\ZulipClient.csproj", "{A5D0D82D-4056-45D7-ADD8-BA541527A2FF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetCore", "NetCore\NetCore.csproj", "{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZulipAPI", "ZulipAPI\ZulipAPI.csproj", "{BAB36F68-2741-49D9-B05F-CA84D17910FA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -42,18 +42,18 @@ Global
 		{A5D0D82D-4056-45D7-ADD8-BA541527A2FF}.Release|x64.Build.0 = Release|Any CPU
 		{A5D0D82D-4056-45D7-ADD8-BA541527A2FF}.Release|x86.ActiveCfg = Release|Any CPU
 		{A5D0D82D-4056-45D7-ADD8-BA541527A2FF}.Release|x86.Build.0 = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|x64.Build.0 = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Debug|x86.Build.0 = Debug|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|x64.ActiveCfg = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|x64.Build.0 = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|x86.ActiveCfg = Release|Any CPU
-		{DE5F81B3-DBC2-49D2-B6C8-BD53D7506675}.Release|x86.Build.0 = Release|Any CPU
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|x64.ActiveCfg = Debug|x64
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|x64.Build.0 = Debug|x64
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|x86.ActiveCfg = Debug|x86
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Debug|x86.Build.0 = Debug|x86
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|x64.ActiveCfg = Release|x64
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|x64.Build.0 = Release|x64
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|x86.ActiveCfg = Release|x86
+		{BAB36F68-2741-49D9-B05F-CA84D17910FA}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
